Allow for configuring a filter, to limti the outputted queues
This is especially handy when using RabbitMQ in a RPC setup. In this case, a lot of short lived queues are created, all following a certain name scheme. This change allows me to provide a filter, so munin doesn't die from the huge amount of rrd files that will eventually accumulate.